vsock: split out Go 1.11+ integration test
